requirements.txt
asttokens==3.0.0
berserk==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.11.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
executing==2.2.0
fonttools==4.56.0
idna==3.10
imutils==0.5.4
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
ndjson==0.3.1
nest-asyncio==1.6.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wrapt==1.17.2
